backend/pyproject.toml
[project]
name = "aikogpt-backend"
version = "1.0.0"
description = "Backend LangGraph pour aikoGPT - Analyse besoins & génération cas d'usage IA"
requires-python = ">=3.11"
dependencies = [
    # LangGraph et LangChain
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    
    # OpenAI
    "openai>=1.0.0",
    
    # Perplexity
    "perplexity-python>=0.1.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    
    # Parsing
    "openpyxl>=3.1.0",  # Excel
    "pypdfium2>=4.0.0",  # PDF
    "PyPDF2>=3.0.0",     # PDF alternatif
    
    # Génération de rapports
    "python-docx>=1.1.0",
    
    # HTTP (pour routes custom si besoin)
    "httpx>=0.27.0",
    
    # Data validation
    "pydantic>=2.0.0",
    
    # Logging et monitoring
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "models", "prompts", "utils", "workflow"]
include = ["graph_factory.py", "main.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

